Move query_omnifocus fields validation into the shared primitive
validateFields lived only in the tool handler, so the resource endpoints
(project.ts, flagged.ts, inbox.ts, today.ts) that call the queryOmnifocus
primitive directly with their own hardcoded field lists bypassed it entirely —
the exact silent-null bug this was meant to close was still reachable through
that path. Moved VALID_FIELDS/validateFields into the primitive itself, right
next to generateFieldMapping which it constrains, so every caller gets the
check for free. The tool handler now just surfaces the primitive's error
through its existing failure path instead of duplicating the check.
